getting-started-with-qiskit/pyproject.toml
[project]
name = "getting-started-with-qiskit"
version = "0.1.0"
description = "IBM Quantum Learning: https://learning.quantum.ibm.com/learning-path/getting-started-with-qiskit"
authors = [
    {name = "0p3r4t0r",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.9, <4"
dependencies = [
    "qiskit[visualization] (>=1.4.2,<2.0.0)",
    "qiskit-ibm-runtime (>=0.37.0,<0.38.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
